pyproject.toml
[project]
name = "airflow-workbook"
version = "0.1.0"
description = ""
authors = [
    {name = "Jackson Faulkner",email = "jackson.faulkner@attentionarc.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "apache-airflow-providers-google>=10.0.0",
    "apache-airflow==2.10.5",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
